pyproject.toml
[tool.poetry]
name = "telegram_bot_wildberries"
version = "0.1.0"
description = ""
authors = ["Анжела <anzhela.namistyuk@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "3.0.0b3"
python-dotenv = "0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
